+++ Integrations/AutofocusV2/AutofocusV2.py
@@ -73,6 +73,55 @@ API_PARAM_DICT = {
         'Malware Family': 'malware_family'
 
     },
+    'search_arguments': {
+        'file_hash': {
+            'api_name': 'alias.hash',
+            'operator': 'contains'
+        },
+        'domain': {
+            'api_name': 'alias.domain',
+            'operator': 'contains'
+        },
+        'ip': {
+            'api_name': 'alias.ip_address',
+            'operator': 'contains'
+        },
+        'url': {
+            'api_name': 'alias.url',
+            'operator': 'contains'
+        },
+        'wildfire_verdict': {
+            'api_name': 'sample.malware',
+            'operator': 'is',
+            'translate': {
+                'Malware': 1,
+                'Grayware': 2,
+                'Benign': 3,
+                'Phishing': 4,
+            }
+        },
+        'first_seen': {
+            'api_name': 'sample.create_date',
+            'operator': 'is in the range'
+        },
+        'last_updated': {
+            'api_name': 'sample.update_date',
+            'operator': 'is in the range'
+        },
+        'time_range': {
+            'api_name': 'session.tstamp',
+            'operator': 'is in the range'
+        },
+        'time_after': {
+            'api_name': 'session.tstamp',
+            'operator': 'is after'
+        },
+        'time_before': {
+            'api_name': 'session.tstamp',
+            'operator': 'is before'
+        }
+    },
+
     'file_indicators': {
         'Size': 'Size',
         'SHA1': 'SHA1',
@@ -194,7 +243,7 @@ def parse_response(resp, err_operation):
             sys.exit(0)
         elif res_json.get("message").find("AF Cookie Not Found") != -1:
             demisto.results(err_msg)
-        elif err_operation == 'Tag details operation failed' and\
+        elif err_operation == 'Tag details operation failed' and \
                 res_json.get("message").find("Tag") != -1 and res_json.get("message").find("not found") != -1:
             demisto.results(err_msg)
             sys.exit(0)
@@ -358,7 +407,8 @@ def parse_coverage_sub_categories(coverage_data):
     for sub_category_name, sub_category_data in coverage_data.items():
         if sub_category_name in SAMPLE_ANALYSIS_COVERAGE_KEYS:
             new_sub_category_data = get_data_from_coverage_sub_category(sub_category_name, sub_category_data)
-            new_sub_category_name = SAMPLE_ANALYSIS_COVERAGE_KEYS.get(sub_category_name).get('display_name')  # type: ignore
+            new_sub_category_name = SAMPLE_ANALYSIS_COVERAGE_KEYS.get(sub_category_name).get(
+                'display_name')  # type: ignore
             new_coverage[new_sub_category_name] = new_sub_category_data
     return {'coverage': new_coverage}
 
@@ -585,6 +635,120 @@ def filter_object_entries_by_dict_values(result_object, response_dict_name):
     return result_object_filtered
 
 
+def search_samples(query=None, scope=None, size=None, sort=None, order=None, file_hash=None, domain=None, ip=None,
+                   url=None, wildfire_verdict=None, first_seen=None, last_updated=None):
+    validate_no_query_and_indicators(query, [file_hash, domain, ip, url, wildfire_verdict, first_seen, last_updated])
+    if not query:
+        validate_no_multiple_indicators_for_search([file_hash, domain, ip, url])
+        query = build_sample_search_query(file_hash, domain, ip, url, wildfire_verdict, first_seen, last_updated)
+    return run_search('samples', query=query, scope=scope, size=size, sort=sort, order=order)
+
+
+def build_sample_search_query(file_hash, domain, ip, url, wildfire_verdict, first_seen, last_updated):
+    indicator_args_for_query = {
+        'file_hash': file_hash,
+        'domain': domain,
+        'ip': ip,
+        'url': url
+    }
+    indicator_list = build_indicator_children_query(indicator_args_for_query)
+    indicator_query = build_logic_query('OR', indicator_list)
+    filtering_args_for_search = {
+        'wildfire_verdict': API_PARAM_DICT['search_arguments'][wildfire_verdict]['translate'][wildfire_verdict],
+        'first_seen': first_seen,
+        'last_updated': last_updated
+    }
+    filters_list = build_children_query(filtering_args_for_search).append(indicator_query)
+    return build_logic_query('AND', filters_list)
+
+
+def search_sessions(query=None, size=None, sort=None, order=None, file_hash=None, domain=None, ip=None, url=None,
+                    time_range=None, time_after=None, time_before=None):
+    validate_no_query_and_indicators(query, [file_hash, domain, ip, url, time_range, time_after, time_before])
+    if not query:
+        validate_no_multiple_indicators_for_search([file_hash, domain, ip, url])
+        query = build_session_search_query(file_hash, domain, ip, url, time_range, time_after, time_before)
+    return run_search('sessions', query=query, size=size, sort=sort, order=order)
+
+
+def build_session_search_query(file_hash, domain, ip, url, time_range, time_after, time_before):
+    indicator_args_for_query = {
+        'file_hash': file_hash,
+        'domain': domain,
+        'ip': ip,
+        'url': url
+    }
+    indicator_list = build_indicator_children_query(indicator_args_for_query)
+    indicator_query = build_logic_query('OR', indicator_list)
+    filtering_args_for_search = {
+        'time_range': time_range,
+        'time_after': time_after,
+        'time_before': time_before
+    }
+    filters_list = build_children_query(filtering_args_for_search).append(indicator_query)
+    return build_logic_query('AND', filters_list)
+
+
+def build_logic_query(logic_operator, condition_list):
+    operator = None
+    if logic_operator == 'AND':
+        operator = 'all'
+    elif logic_operator == 'OR':
+        operator = 'any'
+    return {
+        'operator': operator,
+        'children': condition_list
+    }
+
+
+def build_children_query(args_for_query):
+    children_list = []
+    for key, val in args_for_query.items():
+        field_api_name = API_PARAM_DICT['search_arguments'][key]['api_name']
+        operator = API_PARAM_DICT['search_arguments'][key]['operator']
+        children_list += children_list_generator(field_api_name, operator, [val])
+    return children_list
+
+
+def build_indicator_children_query(args_for_query):
+    for key, val in args_for_query.items():
+        if val:
+            field_api_name = API_PARAM_DICT['search_arguments'][key]['api_name']
+            operator = API_PARAM_DICT['search_arguments'][key]['operator']
+            children_list = children_list_generator(field_api_name, operator, val)
+    return children_list
+
+
+def children_list_generator(field_name, val_list):
+    query_list = []
+    for value in val_list:
+        query_list.append({
+            'field': field_name,
+            'operator': 'contains',
+            'value': value
+        })
+    return query_list
+
+
+def validate_no_query_and_indicators(query, arg_list):
+    if query:
+        for arg in arg_list:
+            if arg:
+                return_error(f'The command can either run a custom query or use the builtin arguments, but not both')
+
+
+def validate_no_multiple_indicators_for_search(arg_list):
+    used_arg = None
+    for arg in arg_list:
+        if arg and used_arg:
+            return_error(f'This command can receive one builtin argument. Two were given: {used_arg}, {arg}')
+        elif arg:
+            used_arg = arg
+    if not used_arg:
+        return_error('In order tp perform a samples/sessions search a query or an indicator must be given.')
+    return
+
+
 ''' COMMANDS'''
 
 
@@ -609,12 +773,21 @@ def test_module():
 
 def search_samples_command():
     args = demisto.args()
+    file_hash = argToList(args.get('file_hash'))
+    domain = argToList(args.get('domain'))
+    ip = argToList(args.get('ip'))
+    url = argToList(args.get('url'))
+    wildfire_verdict = args.get('wildfire_verdict')
+    first_seen = args.get('first_seen')
+    last_updated = args.get('last_updated')
     query = args.get('query')
     scope = args.get('scope')
     max_results = args.get('max_results')
     sort = args.get('sort')
     order = args.get('order')
-    info = run_search('samples', query=query, scope=scope, size=max_results, sort=sort, order=order)
+    info = search_samples(query=query, scope=scope, size=max_results, sort=sort, order=order, file_hash=file_hash,
+                          domain=domain, ip=ip, url=url, wildfire_verdict=wildfire_verdict, first_seen=first_seen,
+                          last_updated=last_updated)
     md = tableToMarkdown(f'Search Samples Info:', info)
     demisto.results({
         'Type': entryTypes['note'],
@@ -627,11 +800,19 @@ def search_samples_command():
 
 def search_sessions_command():
     args = demisto.args()
+    file_hash = argToList(args.get('file_hash'))
+    domain = argToList(args.get('domain'))
+    ip = argToList(args.get('ip'))
+    url = argToList(args.get('url'))
+    time_range = args.get('time_range')
+    time_after = args.get('time_after')
+    time_before = args.get('time_before')
     query = args.get('query')
     max_results = args.get('max_results')
     sort = args.get('sort')
     order = args.get('order')
-    info = run_search('sessions', query=query, size=max_results, sort=sort, order=order)
+    info = search_sessions(query=query, size=max_results, sort=sort, order=order, file_hash=file_hash, domain=domain,
+                           ip=ip, url=url, time_range=time_range, time_after=time_after, time_before=time_before)
     md = tableToMarkdown(f'Search Sessions Info:', info)
     demisto.results({
         'Type': entryTypes['note'],